chore(navbar): restructure navs
