+++ components/sections/navbar/horizontal/links.tsx
@@ -4,7 +4,7 @@ import { MinusSmallIcon } from "@heroicons/react/24/outline";
 import { motion } from "framer-motion";
 import NextLink from "next/link";
 import type { ReactNode } from "react";
-import useActiveSegment from "../../../hooks/use-active-segment";
+import useActiveSegment from "../../../../hooks/use-active-segment";
 
 type Props = {
   children: ReactNode;
