+++ components/sections/navbar/index.tsx
@@ -1,0 +1,9 @@
+import Horizontal from "./horizontal";
+import Vertical from "./vertical";
+
+const Navbar = {
+  Horizontal: Horizontal,
+  Vertical: Vertical,
+};
+
+export default Navbar;
