+++ components/sections/navbar/vertical/link.tsx
@@ -3,7 +3,7 @@
 import NextLink from "next/link";
 import type { ComponentType, Dispatch, SetStateAction } from "react";
 import { motion } from "framer-motion";
-import useActiveSegment from "../../../hooks/use-active-segment";
+import useActiveSegment from "../../../../hooks/use-active-segment";
 
 type Props = {
   id: number;
